// src/utils/types.ts
type Source = {
  "original": string;
  "large2x": string;
  "large": string;
  "medium": string;
  "small": string;
  "portrait": string;
  "landscape": string;
  "tiny": string;
};

export type Photos = {
  "id": number;
  "width": number;
  "height": number;
  "url": string;
  "photographer": string;
  "photographer_url": string;
  "photographer_id": number;
  "avg_color": string;
  "src": Source;
  "liked": boolean;
  "alt": string;
};